+++ features/posts/meeting/domain/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/common/domain/models/MeetingPostDetail.kt
@@ -6,8 +6,8 @@ import kotlin.time.Instant
 data class MeetingPostDetail(
     val post: MeetingPost,
 
-    val maxParticipationCount: Long,
-    val currentParticipantCount: Long,
+    val maxParticipationCount: Int,
+    val currentParticipantCount: Int,
     val isUserJoined: Boolean,
     val recruitmentPeriod: RecruitmentPeriod,
