+++ features/posts/meeting/presentation/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/meeting/presentation/detail/models/MeetingPostDetailUi.kt
@@ -13,8 +13,8 @@ data class MeetingPostDetailUi(
     val post: MeetingPostUi,
     val isLike: Boolean,
 
-    val maxParticipationCount: Long,
-    val currentParticipantCount: Long,
+    val maxParticipationCount: Int,
+    val currentParticipantCount: Int,
     val isUserJoined: Boolean,
     val recruitmentPeriod: RecruitmentPeriodUi,
